Removed App ID info so that users may use their own and actually have it work... ;-)
Search the source for [YOUR APP ID] and replace with your own from the
facebook app dashboard that you have configured. Facebook requires a
secure connection from a particular location on your server, and so
simply installing this app as it was before would not work. Now, once
you do the initial configuration, you may have it function.